code/backend/requirements.txt
python-dotenv==1.0.1
openai==1.26.0
tiktoken==0.5.2
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.37
langchain-openai==0.1.6
langchain-postgres==0.0.3
pypdf==4.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.9
psycopg==3.1.18
psycopg-binary==3.1.18
pgvector==0.2.5
uvicorn==0.29.0
fastapi==0.111.0
pydantic==2.6.4
atlassian-python-api==3.41.7
python-json-logger==2.0.7
jq==1.7.0
unstructured[md]==0.15.0
python-magic==0.4.27
opencv-python==4.10.0.84
nltk==3.8.1
langchain-text-splitters==0.0.1
python-magic==0.4.27
